test(menu): share component mount helpers
